backend/requirements.txt
# Core Flask dependencies
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5

# Environment and configuration
python-dotenv==1.0.0

# LLM integration
openai==1.3.0
requests==2.31.0

# Database
SQLAlchemy==2.0.23

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development tools
# Flask-DebugToolbar removed due to compatibility issues with Flask 2.3.3

